refactor to the getXXXStoreState
