+++ src/store/session/index.ts
@@ -1,2 +1,2 @@
 export type { SessionStore } from './store';
-export { useSessionStore } from './store';
+export { getSessionStoreState,useSessionStore } from './store';
